feat: write insufficient funds subtest
while implementing a new test, first write minimal amount of code for the test
to run and check the failing test output. It's very important to just write
enough code to satisfy compiler. We correct Withdraw method to return error
and for now we have to return something so let's just return nil