feat: update helper for a string to compare against
Errors can be converted to a string with the .Error() method, which I can do in order to compare
it with the string I want now and also making sure that error is not nil to ensure I don't call
.Error() on nil.

t.Fatal which will stop the test if its called. This is because I don't want to make any more
assertions on the error returned if there isn't one around. Without this the test would carry on
to the next step and panic of a nil pointer